requirements.txt
Click==7.0
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.3
jsonify==0.5
MarkupSafe==1.1.1
pydim3==3.0.1
Werkzeug==0.16.0
